Fix Unable to configure jsr356 exception in jetty
