+++ vaadin-spreadsheet-charts/pom.xml
@@ -330,6 +330,9 @@
                     <stopKey>STOP</stopKey>
                     <stopPort>8005</stopPort>
                     <webAppConfig>
+                        <webInfIncludeJarPattern>
+                            .*vaadin.*
+                        </webInfIncludeJarPattern>
                         <contextPath>/</contextPath>
                         <resourceBases>
                             <resourceBase>${project.build.directory}/testwebapp</resourceBase>
@@ -337,7 +340,6 @@
                             <resourceBase>${project.build.outputDirectory}</resourceBase>
                         </resourceBases>
                     </webAppConfig>
-                    <contextXml>${basedir}/src/test/resources/jetty-context.xml</contextXml>
                 </configuration>
             </plugin>
         </plugins>
