+++ vaadin-spreadsheet-charts/src/test/java/com/vaadin/addon/spreadsheet/test/demoapps/SpreadsheetChartsDemoUI.java
@@ -1,0 +1,12 @@
+package com.vaadin.addon.spreadsheet.test.demoapps;
+
+import com.vaadin.annotations.Theme;
+import com.vaadin.annotations.Title;
+import com.vaadin.annotations.Widgetset;
+
+@Theme("demo")
+@Title("Spreadsheet-Charts Integration Demo")
+@SuppressWarnings("serial")
+@Widgetset("com.vaadin.addon.spreadsheet.charts.WidgetSet")
+public class SpreadsheetChartsDemoUI extends SpreadsheetDemoUI {
+}
