+++ vaadin-spreadsheet-charts/src/test/java/com/vaadin/spreadsheet/charts/interactiontests/ScrollTBTest.java
@@ -6,6 +6,7 @@ import org.junit.Test;
 
 import com.vaadin.addon.spreadsheet.elements.SpreadsheetElement;
 import com.vaadin.addon.spreadsheet.test.AbstractSpreadsheetTestCase;
+import com.vaadin.addon.spreadsheet.test.demoapps.SpreadsheetChartsDemoUI;
 
 public class ScrollTBTest extends AbstractSpreadsheetTestCase {
 
@@ -15,5 +16,9 @@ public class ScrollTBTest extends AbstractSpreadsheetTestCase {
         SpreadsheetElement spreadsheetElement = $(SpreadsheetElement.class).first();
         spreadsheetElement.scroll(850);
         compareScreen("overlayScroll");
+    }
+    @Override
+    public Class<?> getUIClass() {
+        return SpreadsheetChartsDemoUI.class;
     }
 }
