--- vaadin-spreadsheet-charts/src/test/resources/jetty-context.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<!DOCTYPE Configure PUBLIC "-//Mort Bay Consulting//DTD Configure//EN" "http://www.eclipse.org/jetty/configure.dtd">
-
-<Configure class="org.eclipse.jetty.webapp.WebAppContext">
-    <Call name="setAttribute">
-        <Arg>org.eclipse.jetty.server.webapp.WebInfIncludeJarPattern</Arg>
-        <Arg>^$</Arg>
-    </Call>
-</Configure>
