+++ vaadin-spreadsheet/pom.xml
@@ -286,11 +286,13 @@
                     <httpConnector>
                         <port>9998</port>
                     </httpConnector>
-                    <useProvidedScope>true</useProvidedScope>
                     <useTestClasspath>true</useTestClasspath>
                     <stopKey>STOP</stopKey>
                     <stopPort>8005</stopPort>
                     <webAppConfig>
+                        <webInfIncludeJarPattern>
+                            .*vaadin.*
+                        </webInfIncludeJarPattern>
                         <contextPath>/</contextPath>
                         <resourceBases>
                             <resourceBase>${project.build.directory}/testwebapp</resourceBase>
@@ -298,7 +300,6 @@
                             <resourceBase>${project.build.outputDirectory}</resourceBase>
                         </resourceBases>
                     </webAppConfig>
-                    <contextXml>${basedir}/src/test/resources/jetty-context.xml</contextXml>
                 </configuration>
             </plugin>
