+++ vaadin-spreadsheet/src/test/java/com/vaadin/addon/spreadsheet/test/demoapps/EmptySpreadsheetUI.java
@@ -2,12 +2,14 @@ package com.vaadin.addon.spreadsheet.test.demoapps;
 
 import com.vaadin.addon.spreadsheet.Spreadsheet;
 import com.vaadin.annotations.Theme;
+import com.vaadin.annotations.Widgetset;
 import com.vaadin.server.VaadinRequest;
 import com.vaadin.ui.UI;
 import com.vaadin.ui.VerticalLayout;
 
 @SuppressWarnings("serial")
 @Theme("demo")
+@Widgetset("com.vaadin.addon.spreadsheet.Widgetset")
 public class EmptySpreadsheetUI extends UI {
