+++ vaadin-spreadsheet/src/test/java/com/vaadin/addon/spreadsheet/test/demoapps/TestUIProviderImpl.java
@@ -1,6 +1,8 @@
 package com.vaadin.addon.spreadsheet.test.demoapps;
 
-import com.vaadin.addon.spreadsheet.test.FreezePaneLocaleUI;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
 import com.vaadin.server.UIClassSelectionEvent;
 import com.vaadin.server.UIProvider;
 import com.vaadin.ui.UI;
@@ -11,27 +13,20 @@ import com.vaadin.ui.UI;
 @SuppressWarnings("serial")
 public class TestUIProviderImpl extends UIProvider {
 
+    private static Logger logger = Logger.getLogger(TestUIProviderImpl.class
+            .getName());
+
     @Override
     public Class<? extends UI> getUIClass(UIClassSelectionEvent event) {
         String name = (event.getRequest()).getPathInfo();
-        if (name.contains(TestexcelsheetUI.class.getSimpleName())) {
-            return TestexcelsheetUI.class;
-        } else if (name.contains(SpreadsheetDemoUI.class.getSimpleName())) {
-            return SpreadsheetDemoUI.class;
-        } else if (name.contains(PushTestCase.class.getSimpleName())) {
-            return PushTestCase.class;
-        } else if (name.contains(TouchUI.class.getSimpleName())) {
-            return TouchUI.class;
-        } else if (name.contains(MultipleSpreadsheetUI.class.getSimpleName())) {
-            return MultipleSpreadsheetUI.class;
-        } else if (name.contains(FreezePaneLocaleUI.class.getSimpleName())) {
-            return FreezePaneLocaleUI.class;
-        } else if (name.contains(TabsheetTestUI.class.getSimpleName())) {
-            return TabsheetTestUI.class;
-        } else if (name.contains(EmptySpreadsheetUI.class.getSimpleName())) {
-            return EmptySpreadsheetUI.class;
-        } else if (name.contains(LayoutSpreadsheetUI.class.getSimpleName())) {
-            return LayoutSpreadsheetUI.class;
+        if (name.startsWith("/")) {
+            name = name.substring(1);
+        }
+        try {
+            String className = name;
+            return Class.forName(className).asSubclass(UI.class);
+        } catch (ClassNotFoundException e) {
+            logger.log(Level.SEVERE, "Could not find UI " + name, e);
         }
         return null;
     }
